refactor: move Cursor source files to src/Cursor subdirectory
