+++ src/CMakeLists.txt
@@ -15,6 +15,7 @@ target_sources(kilo
     PUBLIC
         "${PROJECT_SOURCE_DIR}/src/Terminal/Terminal.hpp"
         "${PROJECT_SOURCE_DIR}/src/TerminalState/TerminalState.hpp"
+        "${PROJECT_SOURCE_DIR}/src/Cursor/Cursor.hpp"
         "${PROJECT_SOURCE_DIR}/includes/Editor.hpp"
         "${PROJECT_SOURCE_DIR}/includes/Utilities.hpp"
         "${PROJECT_SOURCE_DIR}/includes/EditorConfig.hpp"
